// src/contracts/generated/RenderedPreviewPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PageStylePayload } from "./PageStylePayload";
import type { PreviewTextStylePayload } from "./PreviewTextStylePayload";
import type { RenderedPagePayload } from "./RenderedPagePayload";

export type RenderedPreviewPayload = { "mode": "standard", html: string, defaultPageStyle: PageStylePayload, defaultTextStyle: PreviewTextStylePayload, } | { "mode": "a4", pages: Array<RenderedPagePayload>, defaultPageStyle: PageStylePayload, defaultTextStyle: PreviewTextStylePayload, };